+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a test that makes sure all of the plugin's symbols resolve correctly by
+  loading the library with `RTLD_NOW`. This test is only run on Unix-like
+  platforms.
+
 ## [0.2.0] - 2022-01-09
 
 ### Added